requirements.txt
Django==2.2.6
djangorestframework==3.10.3
virtualenv==16.7.5
whitenoise==4.1.4
gunicorn==19.6.0

